requirements.txt
asgiref==3.7.2
Django==5.0.1
django-filter==23.5
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
install==1.3.5
phonenumberslite==8.13.29
pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
